% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateVar_examples.R
\name{bar}
\alias{bar}
\title{Parameters whose descriptions are of the same structure.}
\usage{
bar(xx, yy, zz)
}
\arguments{
\item{xx}{A vector of type numeric}

\item{yy}{A vector of type character}

\item{zz}{A vector of type logical}
}
\description{
The function takes multiple parameters but all the descriptions
follow the same pattern.
}
