% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateVar_examples.R
\name{foo}
\alias{foo}
\title{A fixed parameter with additional text.}
\usage{
foo(x, common_var1, common_var2)
}
\arguments{
\item{x}{A common variable called \code{x}, but some more descriptions depend on the function.\cr Some specific comments of this parameter.}

\item{common_var1}{A common variable which always comes with another variable.}

\item{common_var2}{A common variable which always comes with another variable.}
}
\description{
\code{x} is a common parameter for many functions, but it requires
additional text to be added after the common description.
}
\author{
Shimeng Huang (\email{shimeng.huang@roche.com})
}
