% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateVar_examples.R
\name{baz}
\alias{baz}
\title{Multiple templates.}
\usage{
baz(x, y)
}
\arguments{
\item{x}{A common variable called \code{x}, but some more descriptions depend on the function.\cr Some specific comments of y in \code{baz}.}

\item{y}{A common variable called \code{y}, but some more descriptions depend on the function.\cr Some specific comments of y in \code{baz}.}
}
\description{
The function documentation uses more than one template that
takes an argument and the arguments are even of the same name.
}
\author{
Shimeng Huang (\email{shimeng.huang@roche.com})
}
