% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateVar_examples.R
\name{baz}
\alias{baz}
\title{Parameters whose descriptions are of the same structure.}
\usage{
baz(x, xx, yy)
}
\arguments{
\item{x}{This is a common variable.\cr Some specific comments of this parameter.}

\item{xx}{A vector of type numeric}

\item{yy}{A vector of type character}
}
\description{
The function takes multiple parameters but all the descriptions
follow the same pattern.
}
