% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateVar_examples.R
\name{foo}
\alias{foo}
\title{A fixed parameter with additional text.}
\usage{
foo(x, another_var)
}
\arguments{
\item{x}{This is a common variable.\cr Some specific comments of this parameter.}

\item{another_var}{Some other variable.}
}
\description{
\code{x} is a common parameter for many functions, but it requires
additional text to be added after the common description.
}
