% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateVar_example.R
\name{foo}
\alias{foo}
\title{Some dummy function.}
\usage{
foo(x, another_var)
}
\arguments{
\item{x}{This is a common variable.\cr\cr Some specific comments of this parameter.}

\item{another_var}{Some other variable.}
}
\description{
Some dummy function.
}
